requirements.txt
numpy
pandas
tensorflow
scikit-learn
joblib
pillow
matplotlib
seaborn
opencv-python-headless
streamlit-lottie==0.0.5
requests==2.28.1
streamlit==1.38.0
google-generativeai
streamlit-authenticator
openpyxl
python-dotenv
ydata-profiling
setuptools
st-gsheets-connection
keras
torch
torchvision
fpdf

